TinyPR for Priority class usage
Why ?
Scheduler handles pods as first come , first serve basis in kubernetes.
There is no guarantee that an app with high importance will get scheduled,
when there is a system wide resource crunch.Pod Priority help overcome
this problem by ensuring pods with higher priority gets in front of the
queue for scheduling.

What was added in this change :
- Priority class created for critical and non-critical apps
- Deployments depicting usage for priority class in pods
- Pod disruption budgets and their effects using preemptionPolicy